once: Protect against concurrent access
This isn't really needed in the current implementation, as each remote
server is only contacted from one thread.  But without this, 'once' has
a complicated and brittle interface.  With this, 'once's interface is
simple - it does what it says on the tin.

The unsafePerformIO could be swept under the rug with either the
global-variables or safe-globals modules, if either of them worked.