Make captured blobs visible to unprivileged users
Due to an implementation detail around uid reuse in systemd's DynamicUser
feature, systemd isolates the StateDirectory from the host's filesystem.
We don't want that, so bind-mount it back out.  We can safely do this
while the bejnarkli systemd service is running because the uid won't
change while the service is running.