Remove the blob mount's Before=local-fs.target
By default, all local mounts get a Before=local-fs.target.  This is how
systemd makes sure stuff is mounted before starting services.  Normally
there is this chain of Before/After dependencies:

<all local mount points>
local-fs.target
sysinit.target
<all services>

But in the case of bejnarkli-blobs.mount, we only want this to be
mounted while bejnarkli is running (when we can trust that the
DynamicUser mechanism has set its ownership correctly).  So
bejnarkli-blobs has an After dependency on a service.  This creates a
Before/After cycle.  Systemd breaks cycles by randomly deleting
something in the cycle (ಠ_ಠ), which caused nixos-test.nix to become
flaky.

This change removes the link between bejnarkli-blobs.mount and
local-fs.target.  This is done by setting DefaultDependencies=no, which
then requires adding back the other things that DefaultDependencies sets
(in this case, the deps on unmount.target that ensure that this mount is
unmounted during system shutdown).